+++ src/Registrar/index.js
@@ -1,17 +1,24 @@
 'use strict'
 
+/*
+ * adonis-fold
+ *
+ * (c) Harminder Virk <virk@adonisjs.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+*/
+
 const _ = require('lodash')
 const requireStack = require('require-stack')
 const emitter = new (require('events'))()
-const CE = require('../../src/Exceptions')
+const GE = require('@adonisjs/generic-exceptions')
 const ServiceProvider = require('../../src/ServiceProvider')
 
 /**
  * Registrar class is used to register and boot providers. This
  * should be done once and at the time of booting the app.
  *
- * @module Adonis
- * @submodule fold
  * @class Registrar
  */
 class Registrar {
@@ -96,7 +103,8 @@ class Registrar {
     .map((provider) => {
       const Module = requireStack(provider.trim())
       if (Module.prototype instanceof ServiceProvider === false) {
-        throw CE.RuntimeException.invalidServiceProvider(Module.name)
+        const message = `${Module.name} must extend base service provider class`
+        throw GE.RuntimeException.invoke(message, 500, 'E_INVALID_SERVICE_PROVIDER')
       }
       return new Module(this.Ioc)
     })
@@ -152,7 +160,9 @@ class Registrar {
    */
   providers (arrayOfProviders) {
     if (arrayOfProviders instanceof Array === false) {
-      throw CE.InvalidArgumentException.invalidParameters('register expects an array of providers to be registered')
+      throw GE
+        .InvalidArgumentException
+        .invalidParameter('register expects an array of providers to be registered', arrayOfProviders)
     }
     this._providers = this._getProvidersInstance(arrayOfProviders)
     return this