+++ src/Resolver/index.js
@@ -9,12 +9,7 @@
  * file that was distributed with this source code.
 */
 
-const CE = require('../Exceptions')
-
-/**
- * @module Adonis
- * @submodule fold
- */
+const GE = require('@adonisjs/generic-exceptions')
 
 /**
  * This class will resolve a namespace or a pattern
@@ -59,7 +54,9 @@ class Resolver {
    */
   _validateDirectories (directories) {
     if (!directories || typeof (directories) !== 'object') {
-      throw CE.InvalidArgumentException.invalidParameters('Cannot initiate resolver without registering directories')
+      throw GE
+        .InvalidArgumentException
+        .invalidParameter('Cannot initiate resolver without registering directories', directories)
     }
   }
 
@@ -76,7 +73,9 @@ class Resolver {
    */
   _validateNamespace (appNamespace) {
     if (!appNamespace) {
-      throw CE.InvalidArgumentException.invalidParameters('Cannot initiate resolver without registering appNamespace')
+      throw GE
+        .InvalidArgumentException
+        .invalidParameter('Cannot initiate resolver without registering appNamespace', appNamespace)
     }
   }
 
@@ -95,7 +94,9 @@ class Resolver {
    */
   _makeAppNamespace (binding) {
     if (!this._directories[this._forDirectory]) {
-      throw new Error(`Cannot translate binding, since ${this._forDirectory} is not registered under directories`)
+      throw GE
+        .RuntimeException
+        .invoke(`Cannot translate binding, since ${this._forDirectory} is not registered under directories`)
     }
     const basePath = `${this._appNamespace}/${this._directories[this._forDirectory]}`
     return `${basePath}/${binding.replace(basePath, '')}`
@@ -136,12 +137,25 @@ class Resolver {
    */
   translate (binding) {
     if (typeof (binding) !== 'string') {
-      throw new Error(`Cannot translate ${typeof (binding)}, binding should always be a valid string.`)
+      throw GE
+        .InvalidArgumentException
+        .invalidParameter(`Resolver.translate expects binding to be a valid string`, binding)
     }
 
+    /**
+     * If explicit provider return right away
+     */
     if (binding.startsWith('@provider:')) {
       return this._normalize(binding.replace('@provider:', ''))
     }
+
+    /**
+     * If complete namespace return right away
+     */
+    if (binding.startsWith(`${this._appNamespace}/`)) {
+      return this._normalize(binding)
+    }
+
     return this._forDirectory ? this._normalize(this._makeAppNamespace(binding)) : this._normalize(binding)
   }